+++ src/composites/Iverson.h
@@ -246,6 +246,7 @@ namespace sspo
             MIDI_FEEDBACK_VELOCITY_LOOP_STEP,
             MIDI_FEEDBACK_VELOCITY_INDEX,
             MIDI_FEEDBACK_DIVIDER_SLOW,
+            SET_EUCLIDEAN_HITS_PARAM,
             NUM_PARAMS
         };
         enum InputIds
@@ -292,6 +293,7 @@ namespace sspo
             CLOCK_LIGHT,
             SET_LENGTH_LIGHT,
             MIDI_LEARN_LIGHT,
+            SET_EUCLIDEAN_HITS_LIGHT,
             NUM_LIGHTS
         };
 
@@ -305,6 +307,7 @@ namespace sspo
         bool isSetLength = false;
         bool isClearMapping = false;
         bool isClearAllMapping = false;
+        bool isSetEuclideanHits = false;
         bool clock = false;
         dsp::ClockDivider ledDivider;
 
@@ -315,6 +318,7 @@ namespace sspo
             dsp::TSchmittTrigger<float> learn;
             dsp::TSchmittTrigger<float> reset;
             dsp::TSchmittTrigger<float> clock;
+            dsp::TSchmittTrigger<float> euclideanHits;
 
             std::vector<dsp::TSchmittTrigger<float>> actives;
             std::vector<dsp::TSchmittTrigger<float>> grid;
@@ -370,6 +374,9 @@ namespace sspo
         /// handles when the length button is pressed
         /// to learn the length from the next grid triggered
         void lengthInput();
+
+        void euclideanHitsInput();
+
         /// midi assign mode
         void learnInput();
         /// reset sequencers
@@ -390,6 +397,7 @@ namespace sspo
         {
             gridInputs();
             lengthInput();
+            euclideanHitsInput();
         }
 
         pageChangeInputs();
@@ -463,6 +471,18 @@ namespace sspo
     }
 
     template <class TBase>
+    void IversonComp<TBase>::euclideanHitsInput()
+    {
+        if (triggers.euclideanHits.process (TBase::params[SET_EUCLIDEAN_HITS_PARAM].getValue())
+            && ! isSetLength && ! isLearning)
+        {
+            isSetEuclideanHits = ! isSetEuclideanHits;
+        }
+
+        TBase::lights[SET_EUCLIDEAN_HITS_LIGHT].setBrightness (isSetEuclideanHits);
+    }
+
+    template <class TBase>
     void IversonComp<TBase>::gridInputs()
     {
         for (auto t = 0; t < TRACK_COUNT; ++t)
@@ -476,6 +496,11 @@ namespace sspo
                         //                    tracks[t].setLength (getStepIndex (page, s));
                         TBase::params[LENGTH_1_PARAM + t].setValue (getStepIndex (page, s + 1));
                         isSetLength = false;
+                    }
+                    if (isSetEuclideanHits)
+                    {
+                        tracks[t].setEuclidean (s + 1, tracks[t].getLength());
+                        isSetEuclideanHits = false;
                     }
                     else
                     {
@@ -642,6 +667,9 @@ namespace sspo
                 break;
             case IversonComp<TBase>::MIDI_FEEDBACK_DIVIDER_SLOW:
                 ret = { 0.0f, 1.0f, 0.0f, "Feedback Divider", " ", 0, 1, 0.0f };
+                break;
+            case IversonComp<TBase>::SET_EUCLIDEAN_HITS_PARAM:
+                ret = { 0.0f, 1.0f, 0.0f, "Euclidean_hits", " ", 0, 1, 0.0f };
                 break;
 
             default: